Day 19:
This is the first really interesting day!

The code with r0=1 would take a *long time* to run; the only
way to deal with it is to short-circuit it.

Looking at it, it seems clear how the input for this was
generated: there's a fixed prefix of the input that's always
going to be the same (possibly modulo which registers it
uses), followed by the init code that is jumped to immediately.

The init code then is probably randomly generated to produce
a suitable input for the main algorithm.